bookstore/pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "Rafael do Amaral Costa",email = "rafaelcostatricolor@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-debug-toolbar (>=6.2.0,<7.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
